revert(deploy-status): back out the DSS deploy-branch → tracker-status effort
Reverts the three merged deploy-status PRs to a clean slate per maintainer
decision:
- DSS-1 config schema + env-join resolver (#1967, PR #1977)
- DSS-2 deterministic transition engine (#1968, PR #1980)
- DSS-3 deploy-template GitHub environments (#1969, PR #1983)

Done via `git revert -m 1` of the three merge commits, so the interleaved
non-DSS PRs merged during the effort (#1957 work-item repo-scope, #1960
safety-net guard parity) are untouched. Regenerated plugin copies and the
upstream evidence manifest to match the reverted source.

Preserved: `.lisa/PROJECT_LEARNINGS.md` — the general engineering learnings
captured during the effort are independent of the feature and kept.

Removed: the deploy-status source/tests/rule-pair, the deploy-template
environment declarations, and the plan/roster/cleanup artifacts.

Gates green on the reverted tree: typecheck, lint, lint:slow, 7419 unit
tests, quality-workflow integration.



Work-Item: CodySwannGT/lisa#1986
