+++ plugins/lisa/rules/reference/config-resolution.md
@@ -596,65 +596,6 @@ This is the same `deploy.branches` map already used by env-keyed `done` (reverse
 env from PR base branch) and the build flow (forward: base branch from env);
 `deploy.order` only adds the ranking those two directions never needed.
 
-### Deploy status sync (deployStatusSync)
-
-`deployStatusSync` is an **optional, machine-written** section recorded by the
-deploy-status-sync setup flow. Humans never need to author it; unknown fields
-inside it are preserved on round-trip.
-
-| Field | Type | Default | Meaning |
-| --- | --- | --- | --- |
-| `tier` | string | — | Provisioning tier chosen at setup time |
-| `provisioned` | object (string → string) | — | Provisioned tracker artifact ids, keyed by artifact name |
-| `linearBinding` | `"labels"` \| `"states"` | — | Whether Linear deploy statuses are represented as labels or workflow states |
-| `verifiedAt` | string | — | Strict ISO-8601 UTC timestamp (`Z` suffix, e.g. `2026-01-01T00:00:00Z`) of the last successful verification |
-
-**Ladder resolution.** The environment universe is the keys of
-`deploy.branches`. Each env's done status comes from the tracker's env-keyed
-done map — `github.labels.build.done`, `linear.labels.build.done`, or
-`jira.workflow.done` — with configured entries winning over the built-in
-defaults (`status:on-dev` / `status:on-stg` / `status:done` for label trackers;
-`On Dev` / `On Stg` / `Done` for JIRA). Environments are ordered by
-`deploy.order` when present (its env-name set must exactly match the keys of
-`deploy.branches`); otherwise by the canonical `dev < staging < production`
-rank, and an unrankable custom env name with more than one env configured is a
-config error rather than a guess. A string-valued `done` binds only the
-terminal (highest) rung; lower rungs fall back to the default table. An env
-with a branch but no done status anywhere (possible only for custom names
-outside the default tables) is **skipped**, not an error; an explicitly
-configured done status whose env has no branch **is** a config error. A ladder
-is flagged terminal-only when it resolves to exactly one rung **and** that
-rung's env is the highest-ranked environment of the configured universe
-(post-alias) — a sole surviving lower rung after skips is NOT terminal-only,
-so downstream native closure never fires on an intermediate environment.
-Absent or empty `deploy` yields an empty ladder, never an error.
-
-**Alias.** Iff exactly one of `prod` / `production` appears across the
-configured surfaces (`deploy.branches`, the configured done map,
-`deploy.order`), the two spellings join as one environment — so
-`branches.prod` meets the default `done.production` — and rungs report the
-configured spelling. When both spellings appear they are distinct environments
-and no aliasing occurs. No other alias exists.
-
-**Materialized defaults are fallbacks.** `lisa sync` writes the full default
-done map into config, so a configured done-map entry whose value is EXACTLY
-the default table's value for that env is treated as a materialized fallback,
-not operator intent — and it is invisible for ALL purposes: stripped before
-the env-key union is computed, it never blocks the sole `prod` ↔ `production`
-alias, never raises the done-without-branch error, and leaves the env exactly
-as if it were unconfigured (so `branches.prod` beside a materialized
-`production` entry still aliases to one prod rung). Only a value that DIFFERS
-from the default is deliberate configuration: it participates in the union
-(blocking the alias when it introduces the other spelling) and stays
-protected by the error below.
-
-**Error catalog** (verbatim):
-
-- `Invalid deploy configuration in <source>: a done status ("<status>") is configured for environment "<env>", but deploy.branches has no "<env>" entry. Add deploy.branches.<env> (the git branch that deploys to <env>) or remove "<env>" from the done map.` (custom values only — default-identical entries fall back silently)
-- `Invalid deploy.order in <source>: its environment names must exactly match the keys of deploy.branches. deploy.order has [<order-envs>]; deploy.branches has [<branch-envs>].` (also raised for duplicate `deploy.order` entries)
-- `Invalid deploy.order in <source>: every entry must be an environment name string; found <entry>.`
-- `Invalid deploy configuration in <source>: cannot order environment "<env>". Set deploy.order (environments listed lowest first, e.g. ["dev","staging","production"]) so Lisa knows the promotion order.`
-
 ### What's configurable, what's not
 
 - **Status / label NAMES** are configurable per project — that's the point of the vocabulary maps.
