+++ src/cli/update-check.ts
@@ -67,12 +67,10 @@ export function getDefaultCachePath(cwd: string = process.cwd()): string {
 
 /**
  * Read process.env through one explicit, reviewable exception to the app-template
- * env rule. The CLI needs externally supplied process env for this root opt-out;
- * exported so other CLI modules (e.g. deploy-status-sync) reuse this single
- * exception instead of adding their own.
+ * env rule. The CLI needs externally supplied process env for this root opt-out.
  * @returns Current process environment
  */
-export function getProcessEnv(): NodeJS.ProcessEnv {
+function getProcessEnv(): NodeJS.ProcessEnv {
   // eslint-disable-next-line no-restricted-syntax -- CLI root option must read externally supplied process env once
   return process.env;
 }
