+++ src/core/project-config.ts
@@ -18,11 +18,6 @@ import {
   LEGACY_HARNESS_ALIASES,
   type Harness,
 } from "./config.js";
-import { validateSafeRelativeMarkdownPath } from "./config-field-validation.js";
-import {
-  validateDeployStatusSyncConfig,
-  type DeployStatusSyncConfig,
-} from "./deploy-status-sync.js";
 import {
   validateVerificationConfig,
   type VerificationConfig,
@@ -111,8 +106,6 @@ export interface ProjectConfig {
   readonly learnings?: LearningsConfig;
   /** Optional empirical verification provider configuration. */
   readonly verification?: VerificationConfig;
-  /** Optional machine-written deploy-status-sync provisioning record. */
-  readonly deployStatusSync?: DeployStatusSyncConfig;
 }
 
 /**
@@ -319,16 +312,11 @@ export function validateProjectConfig(
       : validateProjectRulesFile(obj.projectRulesFile, configPath);
   const learnings = validateLearningsConfig(obj.learnings, configPath);
   const verification = validateVerificationConfig(obj.verification, configPath);
-  const deployStatusSync = validateDeployStatusSyncConfig(
-    obj.deployStatusSync,
-    configPath
-  );
   return {
     ...(harness === undefined ? {} : { harness }),
     ...(projectRulesFile === undefined ? {} : { projectRulesFile }),
     ...(learnings === undefined ? {} : { learnings }),
     ...(verification === undefined ? {} : { verification }),
-    ...(deployStatusSync === undefined ? {} : { deployStatusSync }),
   };
 }
 
@@ -360,6 +348,50 @@ function validateOptionalHarness(
   throw new Error(
     `Invalid harness in ${configPath}: expected ${allowed}, got ${JSON.stringify(value)}`
   );
+}
+
+/**
+ * Validate a configurable destination as a safe, relative, non-traversing
+ * Markdown path. Shared by every path-typed config field; the `field` label is
+ * interpolated into each diagnostic so callers get a field-specific message.
+ * @param value - Raw configured path
+ * @param source - Config source for errors
+ * @param field - Config field name used in error messages
+ * @returns Validated project-relative path
+ */
+function validateSafeRelativeMarkdownPath(
+  value: unknown,
+  source: string,
+  field: string
+): string {
+  if (
+    typeof value !== "string" ||
+    value.length === 0 ||
+    value.trim() !== value ||
+    value.includes("\\") ||
+    containsControlCharacter(value) ||
+    /^[a-z]:/iu.test(value) ||
+    path.posix.isAbsolute(value) ||
+    path.win32.isAbsolute(value)
+  ) {
+    throw new Error(
+      `Invalid ${field} in ${source}: expected a safe relative POSIX path`
+    );
+  }
+  const segments = value.split("/");
+  if (
+    segments.some(
+      segment => segment === "" || segment === "." || segment === ".."
+    )
+  ) {
+    throw new Error(
+      `Invalid ${field} in ${source}: path traversal is not allowed`
+    );
+  }
+  if (path.posix.extname(value).toLowerCase() !== ".md") {
+    throw new Error(`Invalid ${field} in ${source}: expected a Markdown file`);
+  }
+  return value;
 }
 
 /**
@@ -447,6 +479,18 @@ function validateLearningsConfig(
     return {};
   }
   return { file: validateLearningsFile(file, source) };
+}
+
+/**
+ * Whether a path contains an ASCII control character.
+ * @param value - Configured path
+ * @returns True when any control character is present
+ */
+function containsControlCharacter(value: string): boolean {
+  return Array.from(value).some(character => {
+    const code = character.charCodeAt(0);
+    return code <= 31 || code === 127;
+  });
 }
 
 /**
