+++ typescript/copy-overwrite/.github/GITHUB_ACTIONS.md
@@ -149,34 +149,6 @@ an unprovisioned gate blocks nothing. Not yet wired for rails
 (`release-rails.yml` doesn't thread approval inputs) or harper-fabric (no
 release.yml call).
 
-**Your deploy job declares a GitHub environment** (deploy-status-sync
-substrate): the deploy job in this project's `deploy.yml` carries an
-`environment:` declaration so every deploy run emits `deployment` /
-`deployment_status` events — the substrate that deploy-status-sync (and
-GitHub-for-Jira) listens to. Stacks with a `determine_environment` job reuse
-its friendly-name resolution
-(`needs.determine_environment.outputs.approval_environment`; unconfigured
-branches degrade to the branch name, same as the approval gate); stacks
-without one use the branch ternary
-`github.ref_name == 'main' && 'production' || github.ref_name`. CDK projects
-own their deploy job — declare the environment on it per the commented
-snippet in the template `deploy.yml`. Exactly one deploy job per workflow
-declares an environment (the job that mutates the target); release/build
-bookkeeping jobs never do.
-
-GitHub environment approvals are **per-job**. If your deploy job named the same
-protected environment as the `release_approval` gate, an approval-gated run
-would prompt you **twice**. To avoid that, on approval-gated runs
-(`determine_environment` sets `require_approval`) the deploy job routes its
-environment to `<env>-deploy` (`format('{0}-deploy', <env>)`) rather than the
-bare env. Leave `<env>-deploy` **unprotected** (no required reviewers): the gate
-fires once at `release_approval`, and the deploy job proceeds without a second
-prompt. A gated run records two deployments — the gate against `<env>` and the
-deploy job against `<env>-deploy` — and the `-deploy` suffix keeps the
-`dev`/`staging`/`production` substring so GitHub-for-Jira still classifies it.
-Ungated runs use the bare env; stacks without a `determine_environment` job
-(rails, harper-fabric) are never gated and need no suffix.
-
 **Blackout Periods** (configurable):
 - Production: No weekends, no late nights (10 PM - 6 AM)
 - Holiday blackouts: Dec 24 - Jan 2, Jul 3-5, Nov 27-29
